Fixed a tun of bugs, and added a bunch more features along the way.
Added a confirmation prompt when pressing Escape during gameplay. Players are now warned for quitting the current game session. This allows them to abort the action if pressed accidentally.
Made it so the post game copy prompt option found in the settings menu also controls the copy prompt found in the pause menu.
Made it so session based achievements don't get coppied along with your actual in-game statistics when using the pause menu or the post game copy prompt after death or victory.
Added a pause menu to the normal and endless game modes. You can resume, view and copy aall of your session-based stats there.