Removed the speaker test complete sound, and added an option in the settings menu to toggle hereing copy and export confirmation sounds.
Added an option in the settings menu to enable or disable copy and export confirmation sounds. This setting controls whether a short chime plays after copying stats or exporting files to and from the clipboard. This is ideal for advanced players who prefer to have less auditory confirmation.
Removed the speaker test complete sound. This sound has been renamed, and is now being used elsewhere.