Fixed bugs, and added some small features related to the alarm system.
Added a voice-based countdown system for the final 10 seconds of the game in level 6. This creates a heightened sense of urgency as players race to reach the exit before time runs out.
Removed the alert sound associated with the boss showing up in level 6. The alarm sound will now serv as the warning indicater for it instead.