Added options to the settings menu to toggle the hereing of countdown and levelup sounds for the game.
Added an option in the settings menu to enable or disable level-up sounds for all game modes. This lets players choose whether they want to hear an audio cue each time they advance a level, offering a quieter experience for those who don't prefer it.
Added an option in the settings menu to enable or disable countdown sounds for the Normal Mode. This gives players the control over whether the dramatic 10-second countdown voice is played during the final phase of level 6. This allows for either tension or silence depending on the user's playstyle.