Enhanced the updater with some much needed features and bug fixes.
Made it so the updater now checks if the update file already exists locally, and skips the download if so.
Added overwrite confirmation prompts after extracting downloaded archives to prevent unintentional file loss.
Introduced automatic extraction of downloaded 7z updates. This is only for Windows users. Mac users are instructed to manually extract the updates, do to automated extraction not being possible on those devices.