You can now enable or disable auto running in the game settings menu, along with other fixed bugs.
Added an option to toggle auto running when the player moves in the game settings menu. The default is holding down the alt or control key to walk, and releasing one of those to run.
Fixed an issue where adjusting the master volume in the sound settings menu didn't work what so ever.