Recoded the weapon system to include the new pistol weapon for the normal mode.
Made it so the pistol weapon is only exclusive to the normal mode. You can press the number 9 key to equip the pistol in normal mode, and machine gun in endless mode. This change also applies to the weapons menu for both modes as well.
Added a new pistol weapon for the normal Mode. Players now have access to this weapon in place of the former machine gun. The machine gun can still be used in endless mode.