Fixed some critical duplication bugs with the translation system for all languages.
Fixed An issue where the check_translations function would append duplicate lines to the language translation file every time it was called, causing clutter and unnecessary repetition.
Made it so the translation system now Preserves complete keys and values without having to aggressivly trimm the contents of the written file.