Added darts as tactical items for the endless game, along with some other small enhancements.
Added the ability to face a certain direction when moving. This is very useful for aiming at various entities when using tactical items in the game.
Added the dart item for purchase in the shop menu. These will stun enemies for a certain amount of time when used from your inventory. This item is only available in the endless game.