Added a new inventory menu, along with some other bug fixes.
Resolved an issue where selecting an item in the inventory menu would cause the wrong item to be used based on the cycle's position while pressing tab to go through it.
Added an inventory menu to make finding and using items a bit easier. You can press the letter I to open it.
Changed the keyboard shortcut from I to K to check how many keys you've collected.