Made some sound conversions to the rest of the sound categories, along with some other enhancements and tweaks.
Updated the readme file with total calculation summeries of sounds found in each section of the sound pack documentation.
Updated the readme file with added documentation for creating sound packs in the game.
Converted the rest of the sound categories to use the new sound pack algorithm in the learn game sounds menu.