Added a bunch of new sounds and enhancements.
Added an option in the game settings menu to enable or disable playing the logo sound upon launching the game.
Added some new sounds to the game. Thanks goes to Keri for the huge contribution. This includes a pritty logo, and a custom wood platform for the toy defender mode.